+++ Classes/DataProviders/OverviewDataProviderInterface.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace YoastSeoForTypo3\YoastSeo\DataProviders;
 
-use Doctrine\DBAL\Driver\ResultStatement;
+use Doctrine\DBAL\Result;
 
 interface OverviewDataProviderInterface
 {
-    public function getResults(array $pageIds = []): ?ResultStatement;
+    public function getResults(array $pageIds = []): ?Result;
 }